vgg19_service/requirements.txt
flask
tensorflow-cpu  
numpy
librosa
opencv-python-headless
soundfile
flask-cors